Tosa cleanup and test work
* Cleaned up a few stragglers from my last merge
* Ported the suites over to Junit4 style suites in Gosu
* Removed typeloader pushing from TestEnv.init() (no longer needed)
* Removed some junk files
