+++ src/tosa/loader/parser/tree/QualifiedAsteriskSelectList.java
@@ -4,22 +4,20 @@ import tosa.loader.parser.Token;
 
 import java.util.Map;
 
-public class QualifiedAsteriskSelectList  extends SQLParsedElement {
-
-  private Token _tableName;
-
+public class QualifiedAsteriskSelectList extends SQLParsedElement {
   public QualifiedAsteriskSelectList(Token start, Token end) {
-    super(start, end);
-    _tableName = start;
+    super(start,  end);
   }
 
   @Override
   protected void toSQL(boolean prettyPrint, int indent, StringBuilder sb, Map<String, Object> values) {
-    sb.append(_tableName.getValue());
+    sb.append("\"");
+    sb.append(getFirst().toString());
+    sb.append("\"");
     sb.append(".*");
   }
 
   public String getTableName() {
-    return _tableName.getValue();
+    return getFirst().toString();
   }
 }
