+++ test/test/query/SampleJoinQueryWithQualifiedAsterisk.sql
