+++ test/test/query/SampleQueryWithComments.sql
