--- test/tosa/loader/DBTypeInfoTestWrapper.java
@@ -1,19 +1,0 @@
-package tosa.loader;
-
-import org.junit.Test;
-import test.GosuTestWrapper;
-
-/**
- * Created by IntelliJ IDEA.
- * User: alan
- * Date: 1/4/11
- * Time: 9:14 PM
- * To change this template use File | Settings | File Templates.
- */
-public class DBTypeInfoTestWrapper extends GosuTestWrapper {
-
-  @Override
-  public String getWrappedTestName() {
-    return "tosa.loader.DBTypeInfoTest";
-  }
-}
