--- tosa-test/test-src/tosa/loader/SQLTypeInfoTestWrapper.java
@@ -1,18 +1,0 @@
-package tosa.loader;
-
-import test.GosuTestWrapper;
-
-/**
- * Created by IntelliJ IDEA.
- * User: alan
- * Date: 1/4/11
- * Time: 9:14 PM
- * To change this template use File | Settings | File Templates.
- */
-public class SQLTypeInfoTestWrapper extends GosuTestWrapper {
-
-  @Override
-  public String getWrappedTestName() {
-    return "tosa.loader.SQLTypeInfoTest";
-  }
-}
